config should emulate other js/ts config files, e.g. jest.config.js
