--- src/editor.config.json
@@ -1,15 +1,0 @@
-{
-    "flow": "a4f64f1b-85bc-477e-b706-de313a022979",
-    "languages": {
-        "eng": "English",
-        "spa": "Spanish"
-    },
-    "endpoints": {
-        "flows": "/assets/flows.json",
-        "groups": "/assets/groups.json",
-        "contacts": "/assets/contacts.json",
-        "fields": "/assets/fields.json",
-        "activity": "",
-        "engine": ""
-    }
-}
