+++ src/services/EditorConfig.ts
@@ -2,6 +2,7 @@ import { ComponentClass, SFC } from 'react';
 import { AnyAction } from '../flowTypes';
 import { Language } from '../components/LanguageSelector';
 import { AnyFormProps } from '../components/NodeEditor/NodeEditor';
+import { languages, Languages, endpoints, Endpoints } from '../editor.config';
 import ChangeGroupComp from '../components/actions/ChangeGroup/ChangeGroup';
 import ChangeGroupForm from '../components/actions/ChangeGroup/ChangeGroupForm';
 import SaveFlowResultComp from '../components/actions/SaveFlowResult/SaveFlowResult';
@@ -18,25 +19,10 @@ import SwitchRouterForm from '../components/routers/SwitchRouter';
 import SubflowRouterForm from '../components/routers/SubflowRouter';
 import WebhookRouterForm from '../components/routers/WebhookRouter';
 
-const { languages, endpoints } = require('../editor.config.json');
-
 export enum Mode {
     EDITING = 0x1,
     TRANSLATING = 0x2,
     ALL = EDITING | TRANSLATING
-}
-
-export interface Endpoints {
-    fields: string;
-    groups: string;
-    engine: string;
-    contacts: string;
-    flows: string;
-    activity: string;
-}
-
-export interface Languages {
-    [iso: string]: string;
 }
 
 export interface Type {
